+++ python-binding/tests/ces/exec.py
@@ -1,45 +1,49 @@
 import json
 
-import asyncio
-from websockets.asyncio.client import connect
+from http import client
 
 import hashlib
 import time
 
 
 def hash_packet(packet) -> str:
-    pkt_time = packet["time"]
-    pkt_type = packet["type"]
-    pkt_content = packet["content"]
-
     hash_str = hashlib.sha256(bytes(
-            str(pkt_time) +
-            str(pkt_type) +
-            str(pkt_content),
+            str(packet["headers"]["p-time"])
+            + str(packet["headers"]["p-type"])
+            + str(packet["content"]),
             "ascii")).hexdigest()
     return hash_str
 
 
 def build_packet(pkt_type, pkt_content):
     packet = {
-        "time": str(time.time()),
-        "type": str(pkt_type),
-        "content": str(pkt_content),
+        "headers": {
+            "p-time": str(time.time()),
+            "p-type": str(pkt_type),
+        },
+        "content": json.dumps(pkt_content),
     }
     hash_str = hash_packet(packet)
-    packet["hash"] = hash_str
+    packet["headers"]["p-hash"] = str(hash_str)
+    packet["headers"]["content-length"] = len(packet["content"])
     return packet
 
 
-def pkt_err(err_type, pkt_content):
-    return build_packet("err:"+err_type, pkt_content)
+def build_packet_static():
+    packet = {
+        "headers": {
+            "p-time": '0',
+            "p-type": 'rpx:ping',
+        },
+        "content": json.dumps({"msg": "pang"}),
+    }
+    hash_str = hash_packet(packet)
+    packet["headers"]["p-hash"] = str(hash_str)
+    packet["headers"]["content-length"] = len(packet["content"])
+    return packet
 
 
-def pkt_confirm(pkt):
-    return build_packet("pkt:recv", str(pkt["hash"]))
-
-
-vfs_string = json.dumps({
+vfs_string = {
     "name": "testvfs",
     "/": {
         "type": 1,
@@ -51,8 +55,8 @@ vfs_string = json.dumps({
         "type": 0,
         "content": "import sys;print('hello world!');print('hello error!',file=sys.stderr)"
     },
-})
-vfs_string_forbidden = json.dumps({
+}
+vfs_string_forbidden = {
     "name": "testvfs",
     "/": {
         "type": 1,
@@ -69,63 +73,82 @@ vfs_string_forbidden = json.dumps({
         "type": 0,
         "content": "import os;print(os.environ);eval('1+1');exec('x=1');__import__('subprocess');x=__import__;x('os');eval=1"
     },
-})
+}
 entry_point = "/main.py"
 
 
-async def send_pkt_ok():
-    async with connect("ws://localhost:56440") as websocket:
-        pkt = build_packet("ces:exec", json.dumps({
-            "vfs": vfs_string,
-            "entryPoint": entry_point,
-        }))
-        await websocket.send(json.dumps(pkt))
-        message = await websocket.recv()
-        print(message)
-        message = await websocket.recv()
-        print(message)
+def test_ok_hash():
+    pkt = build_packet_static()
+    headers = pkt["headers"]
+    content = json.dumps(pkt["content"], separators=(",", ":"), sort_keys=True)
+    conn = client.HTTPConnection("localhost:56440")
+    conn.request("POST", "", content, headers)
+    respose = conn.getresponse()
+    print(respose.status, respose.reason)
+    print(respose.read())
+    conn.close()
 
 
-async def send_pkt_hash_err():
-    async with connect("ws://localhost:56440") as websocket:
-        pkt = build_packet("ces:exec", json.dumps({
-            "vfs": vfs_string,
-            "entryPoint": entry_point,
-        }))
-        pkt["time"] = str(float(pkt["time"])+0.000001)
-        await websocket.send(json.dumps(pkt))
-        message = await websocket.recv()
-        print(message)
+def test_ok():
+    pkt = build_packet("ces:exec", {
+        "vfs": vfs_string, "entryPoint": entry_point})
+    headers = pkt["headers"]
+    content = pkt["content"]
+    content = json.dumps(pkt["content"], separators=(",", ":"), sort_keys=True)
+    conn = client.HTTPConnection("localhost:56440")
+    conn.request("POST", "", content, headers)
+    respose = conn.getresponse()
+    print(respose.status, respose.reason)
+    print(respose.read())
+    conn.close()
 
 
-async def send_pkt_type_err():
-    async with connect("ws://localhost:56440") as websocket:
-        pkt = build_packet("ces:nonexist", json.dumps({
-            "vfs": vfs_string,
-            "entryPoint": entry_point,
-        }))
-        await websocket.send(json.dumps(pkt))
-        message = await websocket.recv()
-        print(message)
-        message = await websocket.recv()
-        print(message)
+def test_hash_err():
+    pkt = build_packet("ces:exec", {
+        "vfs": vfs_string, "entryPoint": entry_point})
+    pkt["headers"]["p-time"] = str(float(pkt["headers"]["p-time"])+0.00001)
+    headers = pkt["headers"]
+    content = pkt["content"]
+    content = json.dumps(pkt["content"], separators=(",", ":"), sort_keys=True)
+    conn = client.HTTPConnection("localhost:56440")
+    conn.request("POST", "", content, headers)
+    respose = conn.getresponse()
+    print(respose.status, respose.reason)
+    print(respose.read())
+    conn.close()
 
 
-async def send_pkt_type_forbid():
-    async with connect("ws://localhost:56440") as websocket:
-        pkt = build_packet("ces:exec", json.dumps({
-            "vfs": vfs_string_forbidden,
-            "entryPoint": entry_point,
-        }))
-        await websocket.send(json.dumps(pkt))
-        message = await websocket.recv()
-        print(message)
-        message = await websocket.recv()
-        print(message)
+def test_type_err():
+    pkt = build_packet("ces:nonexist", {
+        "vfs": vfs_string, "entryPoint": entry_point})
+    headers = pkt["headers"]
+    content = pkt["content"]
+    content = json.dumps(pkt["content"], separators=(",", ":"), sort_keys=True)
+    conn = client.HTTPConnection("localhost:56440")
+    conn.request("POST", "", content, headers)
+    respose = conn.getresponse()
+    print(respose.status, respose.reason)
+    print(respose.read())
+    conn.close()
+
+
+def test_forbid():
+    pkt = build_packet("ces:exec", {
+        "vfs": vfs_string_forbidden, "entryPoint": entry_point})
+    headers = pkt["headers"]
+    content = pkt["content"]
+    content = json.dumps(pkt["content"], separators=(",", ":"), sort_keys=True)
+    conn = client.HTTPConnection("localhost:56440")
+    conn.request("POST", "", content, headers)
+    respose = conn.getresponse()
+    print(respose.status, respose.reason)
+    print(respose.read())
+    conn.close()
 
 
 if __name__ == "__main__":
-    asyncio.run(send_pkt_ok())
-    asyncio.run(send_pkt_hash_err())
-    asyncio.run(send_pkt_type_err())
-    asyncio.run(send_pkt_type_forbid())
+    test_ok()
+    test_hash_err()
+    test_type_err()
+    test_forbid()
+    test_ok_hash()